+++ client.go
@@ -33,6 +33,7 @@ type APIClient struct {
 	Crypto             *Crypto
 	CompanyValuation   *CompanyValuation
 	TechnicalIndicator *TechnicalIndicator
+	API                *API
 	Logger             *zap.Logger
 	Debug              bool
 }
@@ -115,6 +116,7 @@ func NewAPIClient(cfg Config) (*APIClient, error) {
 	APIClient.Crypto = &Crypto{Client: HTTPClient}
 	APIClient.CompanyValuation = &CompanyValuation{Client: HTTPClient}
 	APIClient.TechnicalIndicator = &TechnicalIndicator{Client: HTTPClient}
+	APIClient.API = &API{Client: HTTPClient}
 
 	return APIClient, nil
 }