+++ objects/stock.go
@@ -363,6 +363,17 @@ type HistorySector struct {
 	TechnologyChangesPercentage            float64 `json:"technologyChangesPercentage"`
 }
 
+// SurvivorshipBiasFree ...
+type SurvivorshipBiasFree struct {
+	Symbol string  `json:"symbol"`
+	Open   float64 `json:"open"`
+	High   float64 `json:"high"`
+	Low    float64 `json:"low"`
+	Close  float64 `json:"close"`
+	Volume int64   `json:"volume"`
+	From   string  `json:"from"`
+}
+
 // StockSerieType return string
 func (s StockSerieType) String() string {
 	return string(s)